refactor: move database to the root level
