+++ app/Filament/System/Installer.php
@@ -2,7 +2,6 @@
 
 namespace App\Filament\System;
 
-use App\Database\seeds\ShieldSeeder;
 use App\Filament\Infolists\Components\StreamEntry;
 use App\Models\User;
 use App\Services\AirlineService;
@@ -12,6 +11,7 @@ use App\Services\Installer\SeederService;
 use App\Services\UserService;
 use App\Support\Countries;
 use App\Support\Utils;
+use Database\Seeders\ShieldSeeder;
 use Filament\Actions\Action;
 use Filament\Forms\Components\Select;
 use Filament\Forms\Components\TextInput;
